Feat: update Email Alert
1. 이메일 알림 기능 업데이트합니다.
2. 이메일 알림 비동기 처리에서 다중 엔티티 처리
3. 아티스트 구독 서비스가 미흡하여 보충 후 artist 관련 알림 기능까지 구현하도록 하겠습니다.

+++ src/main/java/com/example/stagealarm/show/dto/ShowInfoRequestDto.java
@@ -23,7 +23,7 @@ public class ShowInfoRequestDto {
     private String price;
 
     @Builder.Default
-    private List<Long> artistIds = new ArrayList<>(); // 아티스트는 여러명일 가능성
+    private List<Integer> artistIds = new ArrayList<>(); // 아티스트는 여러명일 가능성
     @Builder.Default
-    private List<Long> genreIds = new ArrayList<>(); // 공연의 장르는 한가지로 결정.
+    private List<Integer> genreIds = new ArrayList<>(); // 공연의 장르는 한가지로 결정.
 }
